장바구니에서 구매후 장바구니 물품 삭제 안되는 문제 수정

+++ backend/src/main/java/com/example/backend/product/service/CartService.java
@@ -72,9 +72,11 @@ public class CartService {
     }
 
 
-    public void deleteCartItem(Integer memberId, List<CartDeleteRequest> deleteList) {
+    public void deleteCartItem(List<CartDeleteRequest> deleteList) {
         for (CartDeleteRequest req : deleteList) {
-            cartRepository.deleteById(req.getCartId());
+            if (req.getCartId() != null) {
+                cartRepository.deleteById(req.getCartId());
+            }
         }
 
     }
